Added a new modal to let the server owner customize the questions. For now, they will all be required
